Refactor factory pattern
