--- backend/soen343/src/main/java/com/example/soen343/model/Stakeholder.java
@@ -1,16 +1,0 @@
-package com.example.soen343.model;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class Stakeholder extends AbstractUser {
-    private String organizationId;
-
-    public Stakeholder(User user) {
-        super(user.getId(), user.getUsername(), user.getPassword(), user.getType(),
-                user.getEmail(), user.getFullName());
-        this.organizationId = user.getOrganizationId();
-    }
-}
